+++ src-input/duk_util_bufwriter.c
@@ -1,5 +1,5 @@
 /*
- *  Fast buffer writer with spare management.
+ *  Fast buffer writer with slack management.
  */
 
 #include "duk_internal.h"
@@ -57,7 +57,7 @@ DUK_INTERNAL duk_uint8_t *duk_bw_resize(duk_hthread *thr, duk_bufwriter_ctx *bw_
 	 */
 
 	curr_off = (duk_size_t) (bw_ctx->p - bw_ctx->p_base);
-	add_sz = (curr_off >> DUK_BW_SPARE_SHIFT) + DUK_BW_SPARE_ADD;
+	add_sz = (curr_off >> DUK_BW_SLACK_SHIFT) + DUK_BW_SLACK_ADD;
 	new_sz = curr_off + sz + add_sz;
 	if (DUK_UNLIKELY(new_sz < curr_off)) {
 		/* overflow */